requirements.txt
seleniumbase
flask
lxml
sqlalchemy
psycopg2-binary
python-dotenv
gunicorn